+++ src/Modules/Blog/BlogControl.php
@@ -136,7 +136,7 @@ class BlogControl extends Control
 				}
 			}
 
-			$this->pageHelper->addContent($this->view->blog_entry_form($regions, true));
+			$this->pageHelper->addContent($this->view->blog_entry_form($regions));
 
 			$this->pageHelper->addContent($this->v_utils->v_field($this->v_utils->v_menu([
 				$this->routeHelper->pageLink('blog')
@@ -173,10 +173,9 @@ class BlogControl extends Control
 			$this->pageHelper->addBread($this->translator->trans('blog.all'), '/?page=blog&sub=manage');
 			$this->pageHelper->addBread($this->translator->trans('blog.edit'));
 
-			$this->dataHelper->setEditData($data);
 			$regions = $this->session->getRegions();
 
-			$this->pageHelper->addContent($this->view->blog_entry_form($regions));
+			$this->pageHelper->addContent($this->view->blog_entry_form($regions, $data));
 		} else {
 			$this->flashMessageHelper->info($this->translator->trans('blog.permissions.edit'));
 			$this->routeHelper->goPage();